rm calls to logging.basicConfig
fixes #53

The reational is that because richkit is a library, we should not
decide what log messages to print to stdout, but leave that to whatever
application uses richkit. Removed calls to `logging.basicConfig`
bacause that sets up the entire loggin facility and prints to stdout -
also for messages from modules we use.